Separate supabase into auth and database
