+++ pages/signup/mail_sent.tsx
@@ -9,8 +9,8 @@ import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import React, { useEffect } from "react";
 import Footer from "@/components/Footer";
 import Header from "@/components/Header";
+import { useAuth } from "@/hooks/auth";
 import { Checks } from "@/hooks/check"
-import { useAuth } from '@/hooks/supabase'
 
 
 export const getStaticProps: GetStaticProps = async ({ locale }) => ({
